requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<2.0
requests>=2.31,<3.0
pyyaml>=6.0,<7.0
python-dotenv>=1.0,<2.0
openpyxl>=3.1,<4.0
scipy>=1.11,<2.0
pytest>=7.4,<9.0
